move share kid to separate share-kid-service
